pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytincture_example"
version = "0.1.2"
description = "UI Example"
readme = "README.md"
license = { text = "GNU General Public License v2.0" }
authors = [
    { name = "Your Name", email = "schapman1974@gmail.com" }
]
dependencies = [
    "requests==2.32.3",
    "wheel==0.45.1",
    "js==1.0",
    "pyodide-py==0.27.2",
    "pytincture==0.9.0",
    "dhxpyt==0.7.7"
]

[project.urls]
Homepage = "https://github.com/pytincture/pytincture_example"

[tool.setuptools.packages]
find = {}
